+++ coinkiri-application/build.gradle.kts
@@ -7,6 +7,7 @@ bootJar.enabled = false
 jar.enabled = true
 
 dependencies {
+    
     implementation(project(":coinkiri-domain"))
 
     implementation("org.springframework.boot:spring-boot-starter")
