+++ coinkiri-bootstrap/api/build.gradle.kts
@@ -1,6 +1,7 @@
 dependencies {
     implementation(project(":coinkiri-application"))
     implementation(project(":coinkiri-domain"))
+    implementation(project(":persistence"))
 
     implementation("org.springframework.boot:spring-boot-starter-web")
 }
