+++ coinkiri-infrastructure/persistence/src/main/kotlin/com/coinkiri/persistence/adapter/jpa/MemberJpaAdapter.kt
@@ -1,0 +1,17 @@
+package com.coinkiri.persistence.adapter.jpa
+
+import com.coinkiri.application.port.out.jpa.MemberRepository
+import com.coinkiri.domain.member.Member
+import com.coinkiri.persistence.adapter.jpa.entity.MemberJpaEntity
+import com.coinkiri.persistence.adapter.jpa.repository.MemberJpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+class MemberJpaAdapter(
+    private val memberJpaRepository: MemberJpaRepository
+) : MemberRepository {
+
+    override fun save(member: Member): Member {
+        return memberJpaRepository.save(MemberJpaEntity.from(member)).toDomain()
+    }
+}
