+++ coinkiri-infrastructure/persistence/src/main/kotlin/com/coinkiri/persistence/adapter/jpa/repository/MemberJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.coinkiri.persistence.adapter.jpa.repository
+
+import com.coinkiri.persistence.adapter.jpa.entity.MemberJpaEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface MemberJpaRepository : JpaRepository<MemberJpaEntity, Long>
