--- coinkiri-infrastructure/persistence/src/main/kotlin/com/coinkiri/persistence/member/adapter/jpa/repository/MemberJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.coinkiri.persistence.member.adapter.jpa.repository
-
-import com.coinkiri.persistence.member.adapter.jpa.entity.MemberJpaEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface MemberJpaRepository : JpaRepository<MemberJpaEntity, Long>
